final version
Final code for ALPHA. Food_Servo rotates a servo 60 degrees on an hourly basis. Launcher_Motor has two sensors which will spin the motors to launch a tennis ball only if the first sensor senses a ball and the second sensor does not detect anything in front of the launcher.